test google_auth controller, #252
