+++ lib/app_web/controllers/google_auth_controller.ex
@@ -5,11 +5,13 @@ defmodule AppWeb.GoogleAuthController do
   alias App.Ctx.Session
   # alias App.Repo
 
-  def index(conn, %{"code" => code}) do
-    {:ok, token} = ElixirAuthGoogle.get_token(code, conn)
-    {:ok, profile} = ElixirAuthGoogle.get_user_profile(token["access_token"])
+  @elixir_auth_google Application.get_env(:app, :elixir_auth_google) || ElixirAuthGoogle
 
-    person = AppWeb.GoogleAuthController.transform_profile_data_to_person(profile)
+  def index(conn, %{"code" => code}) do
+    {:ok, token} = @elixir_auth_google.get_token(code, conn)
+    {:ok, profile} = @elixir_auth_google.get_user_profile(token["access_token"])
+
+    person = App.Ctx.Person.transform_profile_data_to_person(profile)
 
     # get the person by email
     case App.Ctx.get_person_by_email(profile["email"]) do
@@ -42,45 +44,4 @@ defmodule AppWeb.GoogleAuthController do
         |> render("index.html", person: person)
     end
   end
-
-  @doc """
-  `transform_profile_data_to_person/1` transforms the profile data
-  received from invoking `ElixirAuthGoogle.get_user_profile/1`
-  into a `person` record that can be inserted into the people table.
-
-  ## Example
-
-    iex> transform_profile_data_to_person(%{
-      "email" => "nelson@gmail.com",
-      "email_verified" => true,
-      "family_name" => "Correia",
-      "given_name" => "Nelson",
-      "locale" => "en",
-      "name" => "Nelson Correia",
-      "picture" => "https://lh3.googleusercontent.com/a-/AAuE7mApnYb260YC1JY7a",
-      "sub" => "940732358705212133793"
-    })
-    %{
-      "email" => "nelson@gmail.com",
-      "email_verified" => true,
-      "family_name" => "Correia",
-      "given_name" => "Nelson",
-      "locale" => "en",
-      "name" => "Nelson Correia",
-      "picture" => "https://lh3.googleusercontent.com/a-/AAuE7mApnYb260YC1JY7a",
-      "sub" => "940732358705212133793"
-      "status" => 1,
-      "familyName" => "Correia",
-      "givenName" => "Nelson"
-    }
-  """
-  def transform_profile_data_to_person(profile) do
-    profile
-    |> Map.put("familyName", profile["family_name"])
-    |> Map.put("givenName", profile["given_name"])
-    |> Map.put("locale", profile["locale"])
-    |> Map.put("picture", profile["picture"])
-    |> Map.put("status", 1)
-  end
-
 end
