+++ lib/app_web/controllers/sessions_controller.ex
@@ -1,0 +1,9 @@
+defmodule AppWeb.SessionController do
+  use AppWeb, :controller
+
+  def delete(conn, _params) do
+    conn
+    |> AppWeb.Auth.logout()
+    |> redirect(to: Routes.page_path(conn, :index))
+  end
+end
